django-template/{{ cookiecutter.project_slug }}/requirements.txt
Django>=5.0,<6.0
Dotenv
djangorestframework>=3.15.0
django-cors-headers>=3.14.0
whitenoise
{%- if cookiecutter.use_authentication == "yes" %}
djangorestframework-simplejwt
{%- endif %}
{%- if cookiecutter.use_documentation == "yes" %}
drf-yasg>=1.20.0
{%- endif %}
{%- if cookiecutter.banco_de_dados == "mysql" %}
mysqlclient
{%- endif %}
{%- if cookiecutter.banco_de_dados == "postgresql" %}
psycopg2-binary
{%- endif %}
